fix: fixed styling problems and made separate component for EnhancedTableToolbar
