+++ resources/js/Pages/Characterdetail/CharacterDetailTable.tsx
@@ -8,28 +8,23 @@ import {
     TableContainer,
     TableHead,
     TableRow,
-    Stack,
     IconButton,
     Button,
     TablePagination,
-    Toolbar,
-    Typography,
-    Tooltip,
     Box,
     Checkbox,
-    alpha,
 } from '@mui/material'
 import IosShareIcon from '@mui/icons-material/IosShare'
 import CreateIcon from '@mui/icons-material/Create'
 import DeleteIcon from '@mui/icons-material/Delete'
-import FileDownloadIcon from '@mui/icons-material/FileDownload'
 import {RowData} from './CharacterDetailData'
+import EnhancedTableToolbar from './EnhancedTableToolbar'
 
-const StyledTableCell = styled(TableCell)`
-    padding: '3px 25px';
-`
+const StyledTableCell = styled(TableCell)({
+    padding: '5px 25px',
+})
 
-const ColorButton = styled(Button)`
+const ColorButton = styled(Button)({
     color: '#8da57c',
     borderColor: '#8da57c',
     '&:hover': {
@@ -37,58 +32,7 @@ const ColorButton = styled(Button)`
         borderColor: '#a2bf8e',
         backgroundColor: '#a2bf8e',
     },
-`
-
-const StyledIconButton = styled(Button)`
-    color: 'white';
-`
-
-const EnhancedTableToolbar = ({numSelected}: {numSelected: number}) => (
-    <Toolbar
-        sx={{
-            pl: {sm: 2},
-            pr: {xs: 1, sm: 1},
-            ...(numSelected > 0 && {
-                bgcolor: (theme) =>
-                    alpha(
-                        theme.palette.primary.main,
-                        theme.palette.action.activatedOpacity,
-                    ),
-            }),
-        }}>
-        {numSelected > 0 ? (
-            <Typography
-                sx={{flex: '1 1 100%'}}
-                color='inherit'
-                variant='subtitle1'
-                component='div'>
-                {numSelected} selected
-            </Typography>
-        ) : (
-            <Typography
-                sx={{flex: '1 1 100%'}}
-                variant='h6'
-                id='tableTitle'
-                component='div'>
-                Adventures
-            </Typography>
-        )}
-        {numSelected > 0 ? (
-            <Stack direction='row'>
-                <Tooltip title='Export'>
-                    <IconButton>
-                        <FileDownloadIcon />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title='Delete'>
-                    <IconButton>
-                        <DeleteIcon />
-                    </IconButton>
-                </Tooltip>
-            </Stack>
-        ) : null}
-    </Toolbar>
-)
+})
 
 type CharacterDetailTablePropType = {
     rows: RowData[]
@@ -236,15 +180,21 @@ const CharacterDetailTable = ({rows}: CharacterDetailTablePropType) => {
                                             </ColorButton>
                                         </StyledTableCell>
                                         <StyledTableCell>
-                                            <StyledIconButton aria-label='share'>
+                                            <IconButton
+                                                aria-label='share'
+                                                sx={{color: 'white'}}>
                                                 <IosShareIcon fontSize='small' />
-                                            </StyledIconButton>
-                                            <StyledIconButton aria-label='edit'>
+                                            </IconButton>
+                                            <IconButton
+                                                aria-label='edit'
+                                                sx={{color: 'white'}}>
                                                 <CreateIcon fontSize='small' />
-                                            </StyledIconButton>
-                                            <StyledIconButton aria-label='delete'>
+                                            </IconButton>
+                                            <IconButton
+                                                aria-label='delete'
+                                                sx={{color: 'white'}}>
                                                 <DeleteIcon fontSize='small' />
-                                            </StyledIconButton>
+                                            </IconButton>
                                         </StyledTableCell>
                                     </TableRow>
                                 )
